requirements.txt
peewee==3.17.0
requests==2.31.0
celery==5.3.5
pyinotify
msgpack>=1.0.0
tornado==6.4
redis==4.3.4
retry==0.9.2
pyOpenSSL==25.1.0
pycryptodome==3.23.0
jmespath